requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
appdirs==1.4.4
async-timeout==4.0.1
attrs==21.2.0
base58==2.1.1
bitarray==1.2.2
black==21.11b1
cachetools==4.2.4
cattrs==1.8.0
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.0.3
cytoolz==0.11.2
eth-abi==2.1.1
eth-account==0.5.6
eth-hash==0.3.2
eth-keyfile==0.5.1
eth-keys==0.3.3
eth-rlp==0.2.1
eth-typing==2.2.2
eth-utils==1.10.0
etherscan-python==2.1.0
frozenlist==1.2.0
google-api-core==2.3.2
google-auth==2.3.3
google-cloud-secret-manager==2.8.0
googleapis-common-protos==1.54.0
grpc-google-iam-v1==0.12.3
grpcio==1.42.0
grpcio-status==1.42.0
hexbytes==0.2.2
idna==3.3
ipfshttpclient==0.8.0a2
jsonschema==3.2.0
libcst==0.3.23
lru-dict==1.1.7
multiaddr==0.0.9
multidict==5.2.0
mypy-extensions==0.4.3
netaddr==0.8.0
parsimonious==0.8.1
pathspec==0.9.0
platformdirs==2.4.0
proto-plus==1.19.8
protobuf==3.19.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycryptodome==3.11.0
pydantic==1.8.2
pyrsistent==0.18.0
python-dateutil==2.8.2
PyYAML==6.0
regex==2021.11.10
requests==2.26.0
requests-cache==0.8.1
rlp==2.0.1
rsa==4.8
six==1.16.0
tomli==1.2.2
toolz==0.11.2
typing-inspect==0.7.1
typing_extensions==4.0.0
url-normalize==1.4.3
urllib3==1.26.7
varint==1.0.2
web3==5.24.0
websockets==9.1
yarl==1.7.2
